+++ beacon-chain/rpc/apimiddleware/endpoint_registry_test.go
@@ -1,0 +1,1 @@
+package apimiddleware
